test new dictionary methods.
